requirements.txt
httpx
msgpack-python
numpy
piexif
pillow
pyoxipng
websocket-client
zstd
